Move checkDisabled functions
Move chechDisabled functions from wallets/index.tsx to src/utils/validators/addWallets.ts.
Add javadoc.
